+++ crates/sml-statics/src/error.rs
@@ -54,6 +54,7 @@ pub(crate) enum ErrorKind {
   AppFn,
   Use(Option<str_util::SmolStr>),
   UnreachableHandle,
+  DecWithoutEffect,
 }
 
 struct ErrorKindDisplay<'a> {
@@ -184,6 +185,7 @@ impl fmt::Display for ErrorKindDisplay<'_> {
         f.write_str("brought into scope")
       }
       ErrorKind::UnreachableHandle => f.write_str("unreachable `handle`"),
+      ErrorKind::DecWithoutEffect => f.write_str("declaration with no effect"),
     }
   }
 }
@@ -446,6 +448,7 @@ impl Error {
       ErrorKind::AppFn => Code::n(5037),
       ErrorKind::Use(_) => Code::n(5038),
       ErrorKind::UnreachableHandle => Code::n(5039),
+      ErrorKind::DecWithoutEffect => Code::n(5040),
     }
   }
 
@@ -460,7 +463,8 @@ impl Error {
       | ErrorKind::BoolCase
       | ErrorKind::AppFn
       | ErrorKind::Use(_)
-      | ErrorKind::UnreachableHandle => Severity::Warning,
+      | ErrorKind::UnreachableHandle
+      | ErrorKind::DecWithoutEffect => Severity::Warning,
       _ => Severity::Error,
     }
   }